+++ daterange/mod.ts
@@ -1,2 +1,3 @@
 export * from './date-ranges.ts';
+export * from './types.ts';
 export * from './util.ts';