+++ daterange/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.4] - 2025-08-28
+
+- Added `toJSON` method to `DateRanges` class.
+- Moved type definitions to separate types.ts file.
+- Updated dependencies.
+
 ## [0.2.3] - 2025-08-17
 
 - Moved dependencies to `deno.json`, where they belong.